chore: support large text fields (#55)
The only way to get large fields to display properly would be to use patterns that invoke undefined behavior.

This PR creates a separate variable for these large fields and 64-byte aligns the DisplayContext object (704-bytes).

- add an extended-sized text variable to DisplayCtx.
- change large fields to last display step (ipfs, tranfer, etc).
- rename the UX flow steps to be more concise.
- create secondary UX flow steps for extended text fields.
- change `transaction_t` `vendorFieldLength` to `size_t`.
- add bool (isExtended) to the `setDisplaySteps` method.
- update examples and docs to demonstrate v2 Transfers with vendorFields up to 255 characters.
- update Message example with extended field up to 255 characters.
- update the changelog.

+++ src/operations/transactions/transaction.h
@@ -35,7 +35,7 @@ typedef struct transaction_t {
     uint16_t    type;
     uint8_t     senderPublicKey[PUBLICKEY_COMPRESSED_LEN];
     uint64_t    fee;
-    uint8_t     vendorFieldLength;
+    size_t      vendorFieldLength;
     uint8_t     *vendorField;
     union {
         struct {  // v2
